requirements.txt
afinn==0.1
alabaster==0.7.8
anaconda-client==1.4.0
anaconda-navigator==1.2.1
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
astropy==1.2.1
Babel==2.3.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.10.1
bokeh==0.12.0
boto==2.40.0
Bottleneck==1.1.0
certifi==2021.10.8
cffi==1.6.0
chardet==4.0.0
chest==0.2.3
click==6.6
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.7
conda==4.5.11
conda-build==1.21.3
configobj==5.0.6
contextlib2==0.5.3
cryptography==1.4
cycler==0.10.0
cymem==1.31.2
Cython==0.24
cytoolz==0.9.0.1
dask==0.10.0
datashape==0.5.2
decorator==4.0.10
dill==0.2.8.2
Django==2.2.28
djangorestframework==3.12.4
djangorestframework-simplejwt==4.3.0
docutils==0.12
dynd==0.7.3.dev1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.11.1
Flask-Cors==2.1.2
gevent==1.1.1
greenlet==0.4.10
h5py==2.6.0
HeapDict==1.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.5.1
imagesize==0.7.1
importlib-resources==3.2.1
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
joblib==0.14.1
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==4.1.1
jupyter-core==4.1.0
keras==2.15.0
llvmlite==0.11.0
locket==0.2.0
lxml==3.6.0
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.2
moviepy==1.0.3
mpmath==0.19
msgpack==0.5.6
msgpack-numpy @ file:///tmp/build/80754af9/msgpack-numpy_1605111757211/work
multipledispatch==0.4.8
murmurhash==0.28.0
nb-anacondacloud==1.1.0
nb-conda==1.1.0
nb-conda-kernels==1.0.3
nbconvert==4.2.0
nbformat==4.0.1
nbpresent==3.0.2
networkx==1.11
nltk==3.6.2
nose==1.3.7
notebook==4.2.1
numba==0.26.0
numexpr==2.6.0
numpy==1.18.5
odo==0.5.0
opencv-python==4.4.0.40
openpyxl==2.3.2
pandas==0.24.2
partd==0.3.4
path.py==0.0.0
pathlib2==2.1.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
plac @ file:///tmp/build/80754af9/plac_1608319346096/work
ply==3.8
preshed==1.0.1
proglog==0.1.10
psutil==4.3.0
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
pycosat==0.6.3
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.2.3
Pygments==2.1.3
PyJWT==1.7.1
pyOpenSSL==16.2.0
pyparsing==2.1.4
pytest==2.9.2
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.2.1
QtPy==1.0.2
redis==2.10.5
regex==2022.1.18
requests==2.25.1
rope-py3k==0.9.4.post1
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.1
seaborn==0.8
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
spacy==2.0.12
Sphinx==1.4.1
sphinx-rtd-theme==0.1.9
spyder==2.3.9
SQLAlchemy==1.0.13
sqlparse==0.4.4
statsmodels==0.6.1
sympy==1.0
tables==3.2.2
termcolor==1.1.0
terminado==0.6
textblob==0.17.1
thinc==6.10.3
toolz==0.8.0
tornado==4.3
tqdm==4.64.1
traitlets==4.2.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.26.9
vaderSentiment==3.3.2
Werkzeug==0.11.10
wrapt==1.10.11
xlrd==1.0.0
XlsxWriter==0.9.2
xlwings==0.7.2
xlwt==1.1.2
zipp==1.2.0
